+++ app/clients/dropbox/routes/setup/checkAppFolder.js
