+++ app/clients/dropbox/routes/setup/stages.js
